Prevent client bundle from importing Airtable server utilities
