+++ apps/web/app/lp/investments/[id]/page.tsx
@@ -12,9 +12,8 @@ import {
   XAxis,
   YAxis,
 } from "recharts";
-import type { ExpandedRecord } from "@/lib/airtable";
+import { normalizeFieldKey, type ExpandedRecord } from "@/lib/airtable-shared";
 import { formatCurrencyUSD, formatDate, formatNumber } from "@/lib/format";
-import { normalizeFieldKey } from "@/lib/airtable";
 import { usePolling, type RefreshStatus } from "@/hooks/usePolling";
 
 interface Metrics {
