+++ apps/web/app/lp/page.tsx
@@ -12,10 +12,9 @@ import {
   BarChart,
   Bar,
 } from "recharts";
-import type { ExpandedRecord } from "@/lib/airtable";
+import { normalizeFieldKey, type ExpandedRecord } from "@/lib/airtable-shared";
 import { formatCurrencyUSD, formatDate, formatNumber } from "@/lib/format";
 import { usePolling, type RefreshStatus } from "@/hooks/usePolling";
-import { normalizeFieldKey } from "@/lib/airtable";
 
 interface Metrics {
   commitmentTotal: number;
